+++ core/core/tokens.h
@@ -22,8 +22,8 @@ namespace token
  */
  struct baseToken
 {
-	short startPos; /**< The starting position of an token in the string. @note This index is broken in the current revision*/ 
-	short endPos;  /**< The ending position of an token in the string*/
+	size_t startPos; /**< The starting position of an token in the string. @note This index is broken in the current revision*/ 
+	size_t endPos;  /**< The ending position of an token in the string*/
 	tokenType type; /**< The type of token @see tokenType. */
 
 	virtual tree::nodeDataInterface* node()=0; /**< Constructor for nodeDataInterface object. @return nodeDataInterface pointer to the data coresponding to the type. */
@@ -54,12 +54,12 @@ struct operatorToken : public baseToken, operators::interpreter_operator
 struct parenthesesToken : public baseToken
 {
 
-	short opposit; 
+	size_t opposit; 
 
 	bool isOppening(); /**< @return True if the token represents '(' else false. */
 	bool  hasNode();
 	tree::nodeDataInterface * node();
-	parenthesesToken(short startPos, short endPos);
+	parenthesesToken(size_t startPos, size_t endPos);
 	parenthesesToken();
 
 };
@@ -75,7 +75,7 @@ struct valueToken : public baseToken
 
 	tree::nodeDataInterface* node();
 	bool  hasNode();
-	valueToken(short startPos, short endPos);
+	valueToken(size_t startPos, size_t endPos);
 	valueToken();
 };
 /**
@@ -90,7 +90,7 @@ struct variableToken : public baseToken
 	bool _stack; /**< Stackable. Signifies  if the variable should be put on the stack. Used for the corax virtual machine. */
 	tree::nodeDataInterface* node();
 	bool  hasNode();	
-	variableToken(short startPos, short endPos, memory::memory* mem, math_func::function_interface* func);
+	variableToken(size_t startPos, size_t endPos, memory::memory* mem, math_func::function_interface* func);
 	variableToken();
         variableToken(const variableToken&);
         variableToken operator=(const variableToken&); 
@@ -118,8 +118,8 @@ struct funcToken : public baseToken
 	unsigned short baseWheight; /**< Weight of the operator. Higher weight means higher execution priority.  */
 	tree::nodeDataInterface* node();
 	bool  hasNode();
-	funcToken(short startPos, short endPos, generalFuncPtr ptr);
-        funcToken(short startPos,short endPos, usr_ptr ptr);
+	funcToken(size_t startPos, size_t endPos, generalFuncPtr ptr);
+        funcToken(size_t startPos,size_t endPos, usr_ptr ptr);
 	funcToken();
 private:
             ptr_type_enum ptr_type;
